+++ app_python/src/main.py
@@ -1,30 +1,10 @@
-from datetime import datetime
-
-import pytz
-from fastapi import FastAPI, status
-from fastapi.responses import HTMLResponse
+from fastapi import FastAPI
 
 from .config import settings
+from .routes import router
 
 app = FastAPI(
     title=settings.APP_TITLE,
     version=settings.APP_VERSION,
 )
-
-
-@app.get("/", response_class=HTMLResponse)
-async def index():
-    curr_time = datetime.now(pytz.timezone(settings.TIMEZONE))
-    formatted_time = curr_time.strftime(settings.DATETIME_FORMAT)
-
-    html_content = f"""<!DOCTYPE html>
-<html>
-    <head>
-        <title>{settings.TIMEZONE} Time</title>
-    </head>
-    <body>
-        <h1>Current time in {settings.TIMEZONE}: {formatted_time}</h1>
-    </body>
-</html>"""
-
-    return HTMLResponse(content=html_content, status_code=status.HTTP_200_OK)
+app.include_router(router)
